[수정] 마이페이지 API 가져오기
- count 개수에 있어 int 타입이 아니라 Long 타입으로 변경

+++ src/main/java/com/example/trip/domain/member/dto/GetMyScrapResponse.java
@@ -24,10 +24,10 @@ public class GetMyScrapResponse {
     private String title;
 
     @Schema(description = "게시물 좋아요 개수")
-    private int likes;
+    private Long likes;
 
     @Schema(description = "게시물 스크랩 개수")
-    private int scraps;
+    private Long scraps;
 
     @Schema(description = "게시물 생성 일자")
     private LocalDateTime createdTime;
